requirements.txt
twitchio==1.2.2
python-dotenv==0.15.0
firebase_admin==4.5.3